Move lag estimator for lagd
